Fix data type bug

+++ src/Main.java
@@ -5,7 +5,7 @@ public class Main {
     public static void main(String[] args) {
         short[] indices = Inquirer.inquire();
 
-        int result = Calculator.calculate(indices);
+        float result = Calculator.calculate(indices);
 
         System.out.println("Result: " + result);
     }
